refactor(routes): split server routes into modules
Break routes.rs into smaller modules: auth, library, images,
helpers, and tests; wire them together in routes::build_router.
Remove an empty SseRuntime view artifact in yoink-app/hooks.rs.
Also tidy import formatting in auth middleware tests.
